trying to get a working state
