+++ fv3core/stencils/c_sw.py
@@ -1,15 +1,13 @@
-import dace
 import gt4py.gtscript as gtscript
 from gt4py.gtscript import PARALLEL, computation, interval
 
 import fv3core.utils.gt4py_utils as utils
-from fv3core.decorators import FrozenStencil
+from fv3core.decorators import FrozenStencil, computepath_method
 from fv3core.stencils.d2a2c_vect import DGrid2AGrid2CGridVectors
 from fv3core.utils import corners
 from fv3core.utils.grid import axis_offsets
 from fv3core.utils.typing import FloatField, FloatFieldIJ
 
-from fv3core.utils.gt4py_utils import computepath_method
 
 def geoadjust_ut(
     ut: FloatField,
@@ -743,7 +741,7 @@ class CGridShallowWaterDynamics:
             self.delpc,
             self.ptc,
         )
-        self._D2A2CGrid_Vectors.__call__(uc, vc, u, v, ua, va, ut, vt)
+        self._D2A2CGrid_Vectors(uc, vc, u, v, ua, va, ut, vt)
         if self.namelist.nord > 0:
             self._uf_main(
                 u,
