+++ fv3core/stencils/ray_fast.py
@@ -13,8 +13,7 @@ from gt4py.gtscript import (
 
 import fv3core.utils.global_constants as constants
 import fv3core.utils.gt4py_utils as utils
-from fv3core.utils.gt4py_utils import computepath_method
-from fv3core.decorators import FrozenStencil
+from fv3core.decorators import FrozenStencil, computepath_method
 from fv3core.utils.typing import FloatField, FloatFieldK
