+++ fv3core/stencils/riem_solver_c.py
@@ -6,10 +6,10 @@ from gt4py.gtscript import BACKWARD, FORWARD, PARALLEL, computation, interval, l
 import fv3core._config as spec
 import fv3core.utils.global_constants as constants
 import fv3core.utils.gt4py_utils as utils
-from fv3core.decorators import FrozenStencil
+from fv3core.decorators import FrozenStencil, computepath_method
 from fv3core.stencils.sim1_solver import Sim1Solver
 from fv3core.utils.typing import FloatField, FloatFieldIJ
-from fv3core.utils.gt4py_utils import computepath_method
+
 
 @typing.no_type_check
 def precompute(
