+++ fv3core/stencils/updatedzd.py
@@ -3,11 +3,11 @@ from gt4py.gtscript import BACKWARD, FORWARD, PARALLEL, computation, interval
 
 import fv3core.utils.global_constants as constants
 import fv3core.utils.gt4py_utils as utils
-from fv3core.decorators import FrozenStencil
+from fv3core.decorators import FrozenStencil, computepath_method
 from fv3core.stencils.delnflux import DelnFluxNoSG
 from fv3core.stencils.fvtp2d import FiniteVolumeTransport
 from fv3core.utils.typing import FloatField, FloatFieldIJ, FloatFieldK
-from fv3core.utils.gt4py_utils import computepath_method
+
 
 DZ_MIN = constants.DZ_MIN
 
@@ -326,7 +326,7 @@ class UpdateHeightOnDGrid:
         self._interpolate_to_layer_interface(
             y_area_flux, self._y_area_flux_interface, self._gk, self._beta, self._gamma
         )
-        self.finite_volume_transport.__call__(
+        self.finite_volume_transport(
             height,
             self._crx_interface,
             self._cry_interface,
@@ -339,7 +339,7 @@ class UpdateHeightOnDGrid:
 
         # TODO: in theory, we should check if damp_vt > 1e-5 for each k-level and
         # only compute for k-levels where this is true
-        self.delnflux.__call__(
+        self.delnflux(
             height,
             self._height_x_diffusive_flux,
             self._height_y_diffusive_flux,
