+++ fv3core/stencils/xppm.py
@@ -1,14 +1,13 @@
+import dace
 from gt4py import gtscript
 from gt4py.gtscript import __INLINED, PARALLEL, computation, interval
 
 import fv3core._config as spec
 import fv3core.utils.gt4py_utils as utils
-from fv3core.decorators import FrozenStencil
+from fv3core.decorators import FrozenStencil, computepath_method
 from fv3core.stencils import yppm
 from fv3core.utils.typing import FloatField, FloatFieldIJ
 
-from fv3core.utils.gt4py_utils import computepath_method
-import dace
 
 @gtscript.function
 def final_flux(courant, q, fx1, tmp):
