+++ fv3core/stencils/xtp_u.py
@@ -4,11 +4,10 @@ from gt4py.gtscript import __INLINED, PARALLEL, computation, interval
 
 import fv3core._config as spec
 import fv3core.utils.gt4py_utils as utils
-from fv3core.decorators import FrozenStencil
+from fv3core.decorators import FrozenStencil, computepath_method
 from fv3core.stencils import xppm
 from fv3core.utils.typing import FloatField, FloatFieldIJ
 
-from fv3core.utils.gt4py_utils import computepath_method
 
 @gtscript.function
 def _get_flux(
