+++ fv3core/utils/gt4py_utils.py
@@ -1,4 +1,3 @@
-from dace.frontend.python.common import SDFGConvertible
 import logging
 from functools import wraps
 from typing import Any, Callable, Dict, Hashable, List, Optional, Tuple, Union
@@ -50,6 +49,7 @@ if MPI is not None and MPI.COMM_WORLD.Get_size() > 1:
     gt.config.cache_settings["dir_name"] = ".gt_cache_{:0>6d}".format(
         MPI.COMM_WORLD.Get_rank()
     )
+
 dace.Config.set('default_build_folder', value="{gt_cache}/dacecache".format(
     gt_cache=gt.config.cache_settings["dir_name"]
 ))
@@ -555,96 +555,3 @@ def device_sync() -> None:
     if cp and "cuda" in global_config.get_backend():
         cp.cuda.Device(0).synchronize()
 
-
-class LazyComputepathFunction:
-
-    def __init__(self, func, use_dace):
-        self.func = func
-        self._use_dace = use_dace
-        self.daceprog = dace.program(self.func)
-
-    def __call__(self, *args, **kwargs):
-        if self.use_dace:
-            return self.daceprog(*args, **kwargs)
-        else:
-            return self.func(*args, **kwargs)
-
-    def __sdfg__(self, *args, **kwargs):
-        return self.daceprog.to_sdfg(*args, **self.daceprog.__sdfg_closure__(), **kwargs, save=False)
-
-    def __sdfg_closure__(self, *args, **kwargs):
-        return self.daceprog.__sdfg_closure__(*args, **kwargs)
-
-    def __sdfg_signature__(self):
-        return (self.daceprog.argnames, self.daceprog.constant_args)
-
-    @property
-    def use_dace(self):
-        return self._use_dace or global_config.get_dacemode()
-
-class LazyComputepathMethod:
-
-    bound_callables = dict()
-
-    class SDFGEnabledCallable(SDFGConvertible):
-
-        def __init__(self, lazy_method, obj_to_bind):
-            methodwrapper = dace.method(lazy_method.func)
-            self.obj_to_bind = obj_to_bind
-            self.lazy_method = lazy_method
-            self.daceprog = methodwrapper.__get__(obj_to_bind)
-
-        def __call__(self, *args, **kwargs):
-            if self.lazy_method.use_dace:
-                return self.daceprog.__call__(*args, **kwargs)
-            else:
-                return self.lazy_method.func(self.obj_to_bind, *args, **kwargs)
-
-        def __sdfg__(self, *args, **kwargs):
-            return self.daceprog.to_sdfg(*args, **self.daceprog.__sdfg_closure__(), **kwargs)
-
-        def __sdfg_closure__(self, *args, **kwargs):
-            return self.daceprog.__sdfg_closure__(*args, **kwargs)
-
-        def __sdfg_signature__(self):
-            return (self.daceprog.argnames, self.daceprog.constant_args)
-
-
-    def __init__(self, func, use_dace):
-        self.func = func
-        self._use_dace = use_dace
-
-    def __get__(self, obj, objype=None):
-
-        if (id(obj) , id(self.func)) not in LazyComputepathMethod.bound_callables:
-
-            LazyComputepathMethod.bound_callables[(id(obj) , id(self.func))] = LazyComputepathMethod.SDFGEnabledCallable(self, obj)
-
-        return LazyComputepathMethod.bound_callables[(id(obj), id(self.func))]
-
-    @property
-    def use_dace(self):
-        return self._use_dace or global_config.get_dacemode()
-
-def computepath_method(*args, **kwargs):
-    use_dace = kwargs.get('use_dace', global_config.get_dacemode())
-    def _decorator(method):
-        return LazyComputepathMethod(method, use_dace)
-
-    if len(args)==1 and not kwargs and callable(args[0]):
-        return _decorator(args[0])
-    else:
-        assert not args and len(kwargs)==1 and 'use_dace' in kwargs, "bad args"
-        return _decorator
-
-
-def computepath_function(*args, **kwargs):
-    use_dace = kwargs.get('use_dace', global_config.get_dacemode())
-    def _decorator(function):
-        return LazyComputepathFunction(function, use_dace)
-
-    if len(args)==1 and not kwargs and callable(args[0]):
-        return _decorator(args[0])
-    else:
-        assert not args and len(kwargs)==1 and 'use_dace' in kwargs, "bad args"
-        return _decorator
